Create TableMover class
